Fixed #235 - Modify/Delete Mappings
The import/export icon has been taken from org.eclipse.ui.

+++ org.eclipse.swtchart.extensions/src/org/eclipse/swtchart/extensions/core/MappingsDialog.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Lablicate GmbH.
+ * Copyright (c) 2020, 2021 Lablicate GmbH.
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -12,10 +12,12 @@
  *******************************************************************************/
 package org.eclipse.swtchart.extensions.core;
 
+import java.io.File;
 import java.util.Iterator;
 import java.util.Map;
 
 import org.eclipse.jface.dialogs.Dialog;
+import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.SelectionAdapter;
 import org.eclipse.swt.events.SelectionEvent;
@@ -25,22 +27,37 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
+import org.eclipse.swt.widgets.FileDialog;
+import org.eclipse.swt.widgets.MessageBox;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swtchart.extensions.internal.support.SeriesMapper;
+import org.eclipse.swtchart.extensions.preferences.PreferenceConstants;
 
 public class MappingsDialog extends Dialog {
 
+	private static final String DESCRIPTION = "Mappings";
+	private static final String IMPORT = "Import " + DESCRIPTION;
+	private static final String EXPORT = "Export " + DESCRIPTION;
+	private static final String FILTER_EXTENSION = "*.txt";
+	private static final String FILTER_NAME = "SWTChart Mappings (*.txt)";
+	private static final String FILE_NAME = "SWTChartMappings.txt";
+	//
 	private MappingsListUI mappingsListUI;
+	private ScrollableChart scrollableChart;
+	//
+	private IPreferenceStore preferenceStore = ResourceSupport.getPreferenceStore();
 
-	public MappingsDialog(Shell shell) {
+	public MappingsDialog(Shell shell, ScrollableChart scrollableChart) {
+
 		super(shell);
+		this.scrollableChart = scrollableChart;
 	}
 
 	@Override
 	protected void configureShell(Shell shell) {
 
 		super.configureShell(shell);
-		shell.setText("Setting Mappings");
+		shell.setText(DESCRIPTION);
 	}
 
 	@Override
@@ -59,8 +76,10 @@ public class MappingsDialog extends Dialog {
 	protected Control createDialogArea(Composite parent) {
 
 		Composite composite = (Composite)super.createDialogArea(parent);
+		//
+		createToolbarMain(composite);
 		mappingsListUI = createMappingsList(composite);
-		// createToolbarMain(composite);
+		//
 		updateInput();
 		//
 		return composite;
@@ -73,24 +92,25 @@ public class MappingsDialog extends Dialog {
 		return mappingsListUI;
 	}
 
-	@SuppressWarnings("unused")
 	private void createToolbarMain(Composite parent) {
 
 		Composite composite = new Composite(parent, SWT.NONE);
 		GridData gridData = new GridData(GridData.FILL_HORIZONTAL);
 		gridData.horizontalAlignment = SWT.END;
 		composite.setLayoutData(gridData);
-		composite.setLayout(new GridLayout(2, false));
+		composite.setLayout(new GridLayout(4, false));
 		//
 		createButtonDelete(composite);
 		createButtonDeleteAll(composite);
+		createButtonImport(composite);
+		createButtonExport(composite);
 	}
 
 	private Button createButtonDelete(Composite parent) {
 
 		Button button = new Button(parent, SWT.PUSH);
 		button.setText("");
-		button.setToolTipText("Delete selected mapping(s).");
+		button.setToolTipText("Delete the selected mappings.");
 		button.setImage(ResourceSupport.getImage(ResourceSupport.ICON_DELETE));
 		button.addSelectionListener(new SelectionAdapter() {
 
@@ -98,15 +118,21 @@ public class MappingsDialog extends Dialog {
 			@Override
 			public void widgetSelected(SelectionEvent e) {
 
-				Iterator iterator = mappingsListUI.getStructuredSelection().iterator();
-				while(iterator.hasNext()) {
-					Object object = iterator.next();
-					if(object instanceof Map.Entry) {
-						Map.Entry<String, ISeriesSettings> entry = (Map.Entry<String, ISeriesSettings>)object;
-						SeriesMapper.remove(entry.getKey());
+				MessageBox messageBox = new MessageBox(e.display.getActiveShell(), SWT.ICON_QUESTION | SWT.YES | SWT.NO);
+				messageBox.setText(DESCRIPTION);
+				messageBox.setMessage("Would you like to delete the selected mappings?");
+				int decision = messageBox.open();
+				if(SWT.YES == decision) {
+					Iterator iterator = mappingsListUI.getStructuredSelection().iterator();
+					while(iterator.hasNext()) {
+						Object object = iterator.next();
+						if(object instanceof Map.Entry) {
+							Map.Entry<String, ISeriesSettings> entry = (Map.Entry<String, ISeriesSettings>)object;
+							SeriesMapper.remove(entry.getKey());
+						}
 					}
+					updateInput();
 				}
-				updateInput();
 			}
 		});
 		//
@@ -124,8 +150,94 @@ public class MappingsDialog extends Dialog {
 			@Override
 			public void widgetSelected(SelectionEvent e) {
 
-				SeriesMapper.clear();
-				updateInput();
+				MessageBox messageBox = new MessageBox(e.display.getActiveShell(), SWT.ICON_QUESTION | SWT.YES | SWT.NO);
+				messageBox.setText(DESCRIPTION);
+				messageBox.setMessage("Would you like to delete all mappings?");
+				int decision = messageBox.open();
+				if(SWT.YES == decision) {
+					SeriesMapper.clear();
+					updateInput();
+				}
+			}
+		});
+		//
+		return button;
+	}
+
+	private Button createButtonImport(Composite parent) {
+
+		Button button = new Button(parent, SWT.PUSH);
+		button.setText("");
+		button.setToolTipText(IMPORT);
+		button.setImage(ResourceSupport.getImage(ResourceSupport.ICON_IMPORT));
+		button.addSelectionListener(new SelectionAdapter() {
+
+			@Override
+			public void widgetSelected(SelectionEvent e) {
+
+				FileDialog fileDialog = new FileDialog(e.display.getActiveShell(), SWT.READ_ONLY);
+				fileDialog.setText(IMPORT);
+				fileDialog.setFilterExtensions(new String[]{FILTER_EXTENSION});
+				fileDialog.setFilterNames(new String[]{FILTER_NAME});
+				fileDialog.setFilterPath(preferenceStore.getString(PreferenceConstants.P_PATH_MAPPINGS_IMPORT));
+				String path = fileDialog.open();
+				if(path != null) {
+					preferenceStore.putValue(PreferenceConstants.P_PATH_MAPPINGS_IMPORT, fileDialog.getFilterPath());
+					File file = new File(path);
+					Map<String, ISeriesSettings> mappings = MappingsIO.importSettings(file);
+					for(Map.Entry<String, ISeriesSettings> mapping : mappings.entrySet()) {
+						/*
+						 * Map
+						 */
+						String id = mapping.getKey();
+						ISeriesSettings seriesSettings = null;
+						/*
+						 * Adjust the series
+						 */
+						if(scrollableChart != null) {
+							BaseChart baseChart = scrollableChart.getBaseChart();
+							seriesSettings = baseChart.getSeriesSettings(id);
+							if(seriesSettings != null) {
+								MappingsIO.transferSettings(mapping.getValue(), seriesSettings);
+								MappingsIO.transferSettings(mapping.getValue().getSeriesSettingsHighlight(), seriesSettings.getSeriesSettingsHighlight());
+							}
+						}
+						//
+						SeriesMapper.put(id, seriesSettings == null ? mapping.getValue() : seriesSettings);
+					}
+					//
+					updateInput();
+				}
+			}
+		});
+		//
+		return button;
+	}
+
+	private Button createButtonExport(Composite parent) {
+
+		Button button = new Button(parent, SWT.PUSH);
+		button.setText("");
+		button.setToolTipText(EXPORT);
+		button.setImage(ResourceSupport.getImage(ResourceSupport.ICON_EXPORT));
+		button.addSelectionListener(new SelectionAdapter() {
+
+			@Override
+			public void widgetSelected(SelectionEvent e) {
+
+				FileDialog fileDialog = new FileDialog(e.display.getActiveShell(), SWT.SAVE);
+				fileDialog.setOverwrite(true);
+				fileDialog.setText(EXPORT);
+				fileDialog.setFilterExtensions(new String[]{FILTER_EXTENSION});
+				fileDialog.setFilterNames(new String[]{FILTER_NAME});
+				fileDialog.setFileName(FILE_NAME);
+				fileDialog.setFilterPath(preferenceStore.getString(PreferenceConstants.P_PATH_MAPPINGS_EXPORT));
+				String path = fileDialog.open();
+				if(path != null) {
+					preferenceStore.putValue(PreferenceConstants.P_PATH_MAPPINGS_EXPORT, fileDialog.getFilterPath());
+					File file = new File(path);
+					MappingsIO.exportSettings(file, SeriesMapper.getMappings());
+				}
 			}
 		});
 		//
